+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Verbose output to all public functions for better debugging and monitoring
+
 ## [0.8.0] - 2026-03-28
 
 ### Added